refactor(neon_framework,talk_app): Move rich text rendering to neon_framework
