+++ packages/neon_framework/packages/talk_app/test/message_test.dart
@@ -1,5 +1,4 @@
 import 'package:built_collection/built_collection.dart';
-import 'package:built_value/json_object.dart';
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -23,18 +22,10 @@ import 'package:talk_app/src/widgets/message.dart';
 import 'package:talk_app/src/widgets/reactions.dart';
 import 'package:talk_app/src/widgets/read_indicator.dart';
 import 'package:talk_app/src/widgets/reference_preview.dart';
-import 'package:talk_app/src/widgets/rich_object/deck_card.dart';
-import 'package:talk_app/src/widgets/rich_object/fallback.dart';
-import 'package:talk_app/src/widgets/rich_object/file.dart';
-import 'package:talk_app/src/widgets/rich_object/mention.dart';
 import 'package:timezone/data/latest.dart' as tzdata;
 import 'package:timezone/timezone.dart' as tz;
 
 import 'testing.dart';
-
-class MockOnReferenceClickedCallback extends Mock {
-  void call(String reference);
-}
 
 Widget wrapWidget({
   required Widget child,
@@ -1513,275 +1504,6 @@ void main() {
           expect(find.byType(BottomSheet), findsNothing);
         });
       });
-    });
-  });
-
-  group('buildRichObjectParameter', () {
-    for (final isPreview in [true, false]) {
-      group(isPreview ? 'As preview' : 'Complete', () {
-        group('Mention', () {
-          for (final type in [
-            core.RichObjectParameter_Type.user,
-            core.RichObjectParameter_Type.call,
-            core.RichObjectParameter_Type.guest,
-            core.RichObjectParameter_Type.userGroup,
-          ]) {
-            testWidgets(type.value, (tester) async {
-              final userDetails = MockUserDetails();
-              when(() => userDetails.groups).thenReturn(BuiltList());
-
-              final userDetailsBloc = MockUserDetailsBloc();
-              when(() => userDetailsBloc.userDetails)
-                  .thenAnswer((_) => BehaviorSubject.seeded(Result.success(userDetails)));
-
-              final account = MockAccount();
-
-              await tester.pumpWidgetWithAccessibility(
-                TestApp(
-                  providers: [
-                    Provider<Account>.value(value: account),
-                    NeonProvider<UserDetailsBloc>.value(value: userDetailsBloc),
-                  ],
-                  child: RichText(
-                    text: buildRichObjectParameter(
-                      parameter: core.RichObjectParameter(
-                        (b) => b
-                          ..type = type
-                          ..id = ''
-                          ..name = 'name'
-                          ..iconUrl = '',
-                      ),
-                      textStyle: null,
-                      isPreview: isPreview,
-                    ),
-                  ),
-                ),
-              );
-
-              expect(find.byType(TalkRichObjectMention), isPreview ? findsNothing : findsOne);
-              expect(find.text('name', findRichText: true), findsOne);
-            });
-          }
-        });
-
-        testWidgets('File', (tester) async {
-          final account = MockAccount();
-
-          await tester.pumpWidgetWithAccessibility(
-            TestApp(
-              providers: [
-                Provider<Account>.value(value: account),
-              ],
-              child: RichText(
-                text: buildRichObjectParameter(
-                  parameter: core.RichObjectParameter(
-                    (b) => b
-                      ..type = core.RichObjectParameter_Type.file
-                      ..id = '0'
-                      ..name = 'name',
-                  ),
-                  textStyle: null,
-                  isPreview: isPreview,
-                ),
-              ),
-            ),
-          );
-
-          expect(find.byType(TalkRichObjectFile), isPreview ? findsNothing : findsOne);
-          expect(find.text('name', findRichText: true), findsOne);
-        });
-
-        testWidgets('Deck card', (tester) async {
-          await tester.pumpWidgetWithAccessibility(
-            TestApp(
-              child: RichText(
-                text: buildRichObjectParameter(
-                  parameter: core.RichObjectParameter(
-                    (b) => b
-                      ..type = core.RichObjectParameter_Type.deckCard
-                      ..id = ''
-                      ..name = 'name'
-                      ..boardname = 'boardname'
-                      ..stackname = 'stackname',
-                  ),
-                  textStyle: null,
-                  isPreview: isPreview,
-                ),
-              ),
-            ),
-          );
-
-          expect(find.byType(TalkRichObjectDeckCard), isPreview ? findsNothing : findsOne);
-          expect(find.text('name', findRichText: true), findsOne);
-        });
-
-        testWidgets('Fallback', (tester) async {
-          await tester.pumpWidgetWithAccessibility(
-            TestApp(
-              child: RichText(
-                text: buildRichObjectParameter(
-                  parameter: core.RichObjectParameter(
-                    (b) => b
-                      ..type = core.RichObjectParameter_Type.addressbook
-                      ..id = ''
-                      ..name = 'name',
-                  ),
-                  textStyle: null,
-                  isPreview: isPreview,
-                ),
-              ),
-            ),
-          );
-
-          expect(find.byType(TalkRichObjectFallback), isPreview ? findsNothing : findsOne);
-          expect(find.text('name', findRichText: true), findsOne);
-        });
-      });
-    }
-  });
-
-  group('buildChatMessage', () {
-    test('Preview without newlines', () {
-      final chatMessage = MockChatMessage();
-      when(() => chatMessage.message).thenReturn('123\n456');
-      when(() => chatMessage.messageParameters).thenReturn(BuiltMap());
-
-      var span = buildChatMessage(
-        chatMessage: chatMessage,
-        references: BuiltList(),
-        style: const TextStyle(),
-        onReferenceClicked: (_) {},
-      ).children!.single as TextSpan;
-      expect(span.text, '123\n456');
-
-      span = buildChatMessage(
-        chatMessage: chatMessage,
-        references: BuiltList(),
-        style: const TextStyle(),
-        onReferenceClicked: (_) {},
-        isPreview: true,
-      ).children!.single as TextSpan;
-      expect(span.text, '123 456');
-    });
-
-    group('Unused parameters', () {
-      for (final type in core.RichObjectParameter_Type.values) {
-        test(type, () {
-          final chatMessage = MockChatMessage();
-          when(() => chatMessage.message).thenReturn('test');
-          when(() => chatMessage.messageParameters).thenReturn(
-            BuiltMap({
-              type.value: BuiltMap<String, JsonObject>({
-                'type': JsonObject(type.value),
-                'id': JsonObject(1),
-                'name': JsonObject(''),
-              }),
-            }),
-          );
-
-          final spans = buildChatMessage(
-            chatMessage: chatMessage,
-            references: BuiltList(),
-            style: const TextStyle(),
-            onReferenceClicked: (_) {},
-          ).children!;
-          if (type == core.RichObjectParameter_Type.file) {
-            expect(spans, hasLength(3));
-            expect((spans[0] as WidgetSpan).child, isA<TalkRichObjectFile>());
-            expect((spans[1] as TextSpan).text, '\n');
-            expect((spans[2] as TextSpan).text, 'test');
-          } else {
-            expect((spans.single as TextSpan).text, 'test');
-          }
-        });
-      }
-    });
-
-    test('Used parameters', () {
-      final chatMessage = MockChatMessage();
-      when(() => chatMessage.message).thenReturn('123 {actor1} 456 {actor2} 789');
-      when(() => chatMessage.messageParameters).thenReturn(
-        BuiltMap({
-          'actor1': BuiltMap<String, JsonObject>({
-            'type': JsonObject('user'),
-            'id': JsonObject(''),
-            'name': JsonObject(''),
-          }),
-          'actor2': BuiltMap<String, JsonObject>({
-            'type': JsonObject('user'),
-            'id': JsonObject(''),
-            'name': JsonObject(''),
-          }),
-        }),
-      );
-
-      final spans = buildChatMessage(
-        chatMessage: chatMessage,
-        references: BuiltList(),
-        style: const TextStyle(),
-        onReferenceClicked: (_) {},
-      ).children!;
-      expect(spans, hasLength(5));
-      expect((spans[0] as TextSpan).text, '123 ');
-      expect((spans[1] as WidgetSpan).child, isA<TalkRichObjectMention>());
-      expect((spans[2] as TextSpan).text, ' 456 ');
-      expect((spans[3] as WidgetSpan).child, isA<TalkRichObjectMention>());
-      expect((spans[4] as TextSpan).text, ' 789');
-    });
-
-    test('References', () {
-      final chatMessage = MockChatMessage();
-      when(() => chatMessage.message).thenReturn('a 123 b 456 c');
-      when(() => chatMessage.messageParameters).thenReturn(BuiltMap());
-
-      final callback = MockOnReferenceClickedCallback();
-
-      final spans = buildChatMessage(
-        chatMessage: chatMessage,
-        references: BuiltList(['123', '456']),
-        style: const TextStyle(),
-        onReferenceClicked: callback.call,
-      ).children!;
-      expect(spans, hasLength(5));
-
-      expect((spans[0] as TextSpan).text, 'a ');
-
-      final link1 = spans[1] as TextSpan;
-      expect(link1.text, '123');
-      (link1.recognizer! as TapGestureRecognizer).onTap!();
-      verify(() => callback('123')).called(1);
-
-      expect((spans[2] as TextSpan).text, ' b ');
-
-      final link2 = spans[3] as TextSpan;
-      expect(link2.text, '456');
-      (link2.recognizer! as TapGestureRecognizer).onTap!();
-      verify(() => callback('456')).called(1);
-
-      expect((spans[4] as TextSpan).text, ' c');
-    });
-
-    test('Skip empty parts', () {
-      final chatMessage = MockChatMessage();
-      when(() => chatMessage.message).thenReturn('{actor}');
-      when(() => chatMessage.messageParameters).thenReturn(
-        BuiltMap({
-          'actor': BuiltMap<String, JsonObject>({
-            'type': JsonObject(core.RichObjectParameter_Type.user.name),
-            'id': JsonObject(''),
-            'name': JsonObject(''),
-          }),
-        }),
-      );
-
-      final spans = buildChatMessage(
-        chatMessage: chatMessage,
-        references: BuiltList(),
-        style: const TextStyle(),
-        onReferenceClicked: (_) {},
-      ).children!;
-      expect(spans, hasLength(1));
-      expect((spans[0] as WidgetSpan).child, isA<TalkRichObjectMention>());
     });
   });
 }
