+++ packages/neon_framework/test/goldens/rich_text_object_deck_card.png
