+++ packages/neon_framework/test/goldens/rich_text_object_fallback_with_icon.png
