+++ packages/neon_framework/test/goldens/rich_text_object_mention_user_other.png
